Split SoaLocalizedBasisSet
