+++ src/QMCWaveFunctions/CMakeLists.txt
@@ -91,6 +91,8 @@ IF(OHMMS_DIM MATCHES 3)
       LCAO/LCAOrbitalBuilder.cpp
       LCAO/SoaSphericalTensor.cpp
       LCAO/MultiQuinticSpline1D.cpp
+      LCAO/AOBasisBuilder.cpp
+      LCAO/SoaLocalizedBasisSet.cpp
       )
   IF(QMC_COMPLEX)
     SET(FERMION_SRCS ${FERMION_SRCS} LCAO/LCAOSpinorBuilder.cpp)
